--- src/api/auth/login.ts
@@ -1,8 +1,0 @@
-import type { RequestLoginType, ResponseLoginType } from '@/types/api/login';
-import api from '@/utils/api';
-
-export const login = async (
-  rq: RequestLoginType,
-): Promise<ResponseLoginType> => {
-  return await api.post<RequestLoginType, ResponseLoginType>('login', rq);
-};
