--- src/api/auth/signup.ts
@@ -1,8 +1,0 @@
-import type { RequestSignupType, ResponseSignupType } from '@/types/api/signup';
-import api from '@/utils/api';
-
-export const signup = async (
-  rq: RequestSignupType,
-): Promise<ResponseSignupType> => {
-  return await api.post<RequestSignupType, ResponseSignupType>('signup', rq);
-};
