--- src/api/settings/updatePassword/index.ts
@@ -1,8 +1,0 @@
-import api from '@/utils/api';
-import type { RequestLoginType, ResponseLoginType } from '@/types/api/login';
-
-export const login = async (
-  rq: RequestLoginType,
-): Promise<ResponseLoginType> => {
-  return await api.post<RequestLoginType, ResponseLoginType>('login', rq);
-};
