+++ src/components/signup/useSignupForm.ts
@@ -1,6 +1,6 @@
-import { signup } from '@/api/auth/signup';
+import { signup } from '@/api/auth';
+import { ResponseSignupType } from '@/types/api/auth';
 import useForm, { FormErrors, FormValues } from '@/hooks/useForm';
-import { ResponseSignupType } from '@/types/api/signup';
 import validation from '@/utils/validations';
 
 const initialSignupFormValues = {
