--- src/types/api/login/RequestLoginType.ts
@@ -1,4 +1,0 @@
-export interface RequestLoginType {
-  email: string;
-  password: string;
-}
