--- src/types/api/login/ResponseLoginType.ts
@@ -1,5 +1,0 @@
-import type { UserType } from '@/types';
-export interface ResponseLoginType {
-  user: UserType;
-  token: string;
-}
