--- src/types/api/login/index.ts
@@ -1,2 +1,0 @@
-export type { RequestLoginType } from './RequestLoginType';
-export type { ResponseLoginType } from './ResponseLoginType';
