--- src/types/api/signup/RequestSignupType.ts
@@ -1,5 +1,0 @@
-export interface RequestSignupType {
-  email: string;
-  fullName: string;
-  password: string;
-}
