--- src/types/api/signup/ResponseSignupType.ts
@@ -1,5 +1,0 @@
-import type { UserType } from '@/types';
-export interface ResponseSignupType {
-  user: UserType;
-  token: string;
-}
