--- src/types/api/signup/index.ts
@@ -1,2 +1,0 @@
-export type { RequestSignupType } from './RequestSignupType';
-export type { ResponseSignupType } from './ResponseSignupType';
